+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * Uses new version of engineyard-serverside 2.0.7 which fixes an encoding error that may occur.
 
 ## v2.0.12 (2013-04-11)